cleanup trace levels; fix typecasts, typos and formatting issues

+++ crypto/cmp/cmp_int.h
@@ -720,7 +720,7 @@ DECLARE_ASN1_FUNCTIONS(CMP_PROTECTEDPART)
 int CMP_ASN1_OCTET_STRING_set1(ASN1_OCTET_STRING **tgt,
                                const ASN1_OCTET_STRING *src);
 int CMP_ASN1_OCTET_STRING_set1_bytes(ASN1_OCTET_STRING **tgt,
-                                     const unsigned char *bytes, size_t len);
+                                     const unsigned char *bytes, int len);
 X509_EXTENSIONS *CMP_X509_EXTENSIONS_dup(const X509_EXTENSIONS *exts);
 /* from cmp_ctx.c */
 ASN1_OCTET_STRING *CMP_CTX_get0_last_senderNonce(const OSSL_CMP_CTX *ctx);
